백지영: getPharmacyList 연습 종료 10월 22일 16시 44분
